Refactor string newtype duplication
