+++ crates/runtime/src/lib.rs
@@ -2,8 +2,11 @@ mod dialogue;
 mod dialogue_option;
 mod handlers;
 mod line;
+mod string_newtype;
 mod variable_storage;
 mod virtual_machine;
+
+pub(crate) use string_newtype::string_newtype;
 
 pub mod prelude {
     pub(crate) use crate::virtual_machine::*;
